% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{associations}
\alias{associations}
\title{Get Associations by SNP ID and Study ID}
\usage{
associations(snp_ids, study_ids)
}
\arguments{
\item{snp_ids}{A vector of numeric values specifying the SNP IDs}

\item{study_ids}{A vector of numeric values specifying the Study IDs}
}
\value{
A dataframe containing the associations
}
\description{
Get associations from the API by SNP id and study id
}
\section{associations_dataframe}{

The associations dataframe contains information about which studies have association results.
It has the following columns:
\itemize{
\item snp_id: the id of the SNP associated with this association
\item study_id: the id of the study associated with this association
\item beta: the beta value of the association
\item se: the standard error of the association
\item p: the p-value of the association
\item eaf: the estimated allele frequency of the association
\item imputed: whether the association is imputed
}
}

