% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{associations_api}
\alias{associations_api}
\title{Get Associations by SNP ID and Study ID API}
\usage{
associations_api(snp_ids, study_ids)
}
\arguments{
\item{snp_ids}{A vector of numeric values specifying the SNP IDs}

\item{study_ids}{A vector of numeric values specifying the Study IDs}
}
\value{
A list containing the associations
}
\description{
Get associations from the API by SNP id and study id
}
