% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{associations_doc}
\alias{associations_doc}
\title{Associations Dataframe Documentation}
\usage{
associations_doc()
}
\description{
Shared documentation for associations dataframe columns
}
\section{associations dataframe}{

The associations dataframe contains information about which studies have association results.
It has the following columns:
\itemize{
\item snp_id: the id of the SNP associated with this association
\item study_id: the id of the study associated with this association
\item beta: the beta value of the association
\item se: the standard error of the association
\item p: the p-value of the association
\item eaf: the estimated allele frequency of the association
\item imputed: whether the association is imputed
}
}

\keyword{internal}
