% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{cleanup_api_object}
\alias{cleanup_api_object}
\title{Clean up an api object}
\usage{
cleanup_api_object(api_object)
}
\arguments{
\item{api_object}{A list of api info, as returned by the api}
}
\value{
An updated list of api info with null, empty, and NA elements removed
}
\description{
Clean up an api object
}
