% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{coloc_groups_doc}
\alias{coloc_groups_doc}
\title{Coloc Groups Dataframe Documentation}
\usage{
coloc_groups_doc()
}
\description{
Shared documentation for coloc_groups dataframe columns
}
\section{coloc_groups_dataframe}{

The coloc_groups dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item coloc_group_id: the unique id for this group of colocalised results
\item study_id: the id of the study
\item study_extraction_id: the id of the study extraction
\item snp_id: the id of the SNP
\item ld_block_id: the id of the LD block
\item group_threshold: the threshold for this group of colocalised results.
'strong' signifies h4 > 0.8, 'moderate' signifies h4 > 0.6
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item cis_trans: the cis/trans status of the SNP
\item ld_block: the LD block of the SNP
\item display_snp: the display SNP name
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
}
}

\keyword{internal}
