% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{coloc_pairs_doc}
\alias{coloc_pairs_doc}
\title{Coloc Pairs Dataframe Documentation}
\usage{
coloc_pairs_doc()
}
\description{
Shared documentation for coloc_pairs dataframe columns
}
\section{coloc_pairs dataframe}{

The coloc_pairs dataframe contains information about which studies have coloc pairs.
It has the following columns:
\itemize{
\item study_extraction_a_id: the id of the study extraction associated with this coloc pair
\item study_extraction_b_id: the id of the study extraction associated with this coloc pair
\item ld_block_id: the id of the LD block
\item h3: the h3 value for this coloc pair
\item h4: the h4 value for this coloc pair
\item spurious: whether this coloc pair is spurious
}
}

\keyword{internal}
