% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{combine_genes}
\alias{combine_genes}
\title{Combine genes}
\usage{
combine_genes(
  gene_ids,
  include_associations = FALSE,
  include_coloc_pairs = FALSE,
  include_trans = TRUE,
  h4_threshold = 0.8
)
}
\arguments{
\item{gene_ids}{A vector of gene ids}

\item{include_associations}{A logical value specifying whether to include associations
(BETA, SE, P), defaults to FALSE}

\item{include_coloc_pairs}{A logical value specifying whether to include coloc pairs, defaults to FALSE}

\item{include_trans}{A logical value specifying whether to include trans genetic effects, defaults to TRUE}

\item{h4_threshold}{A numeric value specifying the h4 threshold for coloc pairs, defaults to 0.8}
}
\value{
A list which contains the following elements:
\itemize{
\item genes: A list of genes keyed by gene id
}
}
\description{
Combine genes into a single object, merging sub-lists and deduplicating results.
}
