% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{create_source_url}
\alias{create_source_url}
\title{Create a source url for a study}
\usage{
create_source_url(study_name)
}
\arguments{
\item{study_name}{character string specifying the study name}
}
\value{
A character string specifying the source url
}
\description{
Create a source url for a study
}
