% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{gene}
\alias{gene}
\title{Get a Gene}
\usage{
gene(gene_id, include_associations = FALSE)
}
\arguments{
\item{gene_id}{A character string specifying the gene ID}

\item{include_associations}{A logical value specifying whether to include associations}
}
\value{
A list containing the gene information
}
\description{
Get a gene from the API
}
