% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{gene_pleiotropies_api}
\alias{gene_pleiotropies_api}
\title{Get All Gene Pleiotropies API}
\usage{
gene_pleiotropies_api()
}
\value{
A list containing the gene pleiotropies
}
\description{
Get all gene pleiotropies from the API
}
