% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{gene_pleiotropies_doc}
\alias{gene_pleiotropies_doc}
\title{Gene Pleiotropies Dataframe Documentation}
\usage{
gene_pleiotropies_doc()
}
\description{
Shared documentation for gene_pleiotropies dataframe columns
}
\section{gene_pleiotropies_dataframe}{

The gene_pleiotropies dataframe contains information about which genes are pleiotropic.
It has the following columns:
\itemize{
\item gene_id: the id of the gene
\item gene: the name of the gene
\item : distinct_trait_categories the number of trait categories that the gene is associated with via coloc groups
\item : distinct_protein_coding_genes the number of genes that the gene is associated with via coloc groups
}
}

\keyword{internal}
