% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{genes_in_region_doc}
\alias{genes_in_region_doc}
\title{Genes in Region Dataframe Documentation}
\usage{
genes_in_region_doc()
}
\description{
Shared documentation for genes_in_region dataframe columns
}
\section{genes_in_region_dataframe}{

The genes_in_region dataframe contains information about which genes are in a region.
It has the following columns:
\itemize{
\item id: the id of the gene
\item ensembl_id: the ensembl id of the gene
\item gene: the name of the gene
\item description: the description of the gene
\item gene_biotype: the gene biotype
\item chr: the chromosome of the gene
\item start: the start position of the gene
\item stop: the stop position of the gene
\item strand: the strand of the gene
\item source: the source of the gene
}
}

\keyword{internal}
