% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiotropy.R
\name{get_all_snp_pleiotropies}
\alias{get_all_snp_pleiotropies}
\title{Get All SNP Pleiotropies}
\usage{
get_all_snp_pleiotropies()
}
\value{
A list containing the SNP pleiotropies
}
\description{
Get all SNP pleiotropies from the API
}
