% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiotropy.R
\name{get_all_variant_pleiotropies}
\alias{get_all_variant_pleiotropies}
\title{Get All SNP Pleiotropies}
\usage{
get_all_variant_pleiotropies()
}
\value{
A list containing the SNP pleiotropies
\itemize{
\item variant_id: the id of the SNP
\item display_snp: the name of the SNP
\item distinct_trait_categories: the number of trait categories that the SNP is associated with via coloc groups
\item distinct_protein_coding_genes: the number of genes that the SNP is associated with via coloc groups
}
}
\description{
Get all SNP pleiotropies from the API
}
