% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{get_gwas}
\alias{get_gwas}
\title{Get a GWAS from the API}
\usage{
get_gwas(gwas_id, include_summary_stats = FALSE)
}
\arguments{
\item{gwas_id}{The ID of the GWAS}

\item{include_summary_stats}{Whether to include summary statistics}
}
\value{
A list containing the GWAS information
}
\description{
Get a GWAS from the API
}
