% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{get_variants_with_options}
\alias{get_variants_with_options}
\title{Get Variants with Options}
\usage{
get_variants_with_options(
  url,
  include_associations = FALSE,
  p_value_threshold = NULL
)
}
\arguments{
\item{url}{A character string specifying the URL}

\item{include_associations}{A logical value specifying whether to include associations}

\item{p_value_threshold}{A numeric value specifying the p-value threshold}
}
\value{
A list containing the variants
}
\description{
Get variants from the API with options
}
