% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{ld_doc}
\alias{ld_doc}
\title{LD Dataframe Documentation}
\usage{
ld_doc()
}
\description{
Shared documentation for ld dataframe columns
}
\section{ld_dataframe}{

The ld dataframe contains information about the LD matrix.
It has the following columns:
\itemize{
\item lead_snp_id: the id of the lead SNP
\item variant_snp_id: the id of the variant SNP
\item ld_block_id: the id of the LD block
\item r: the r value between the lead and variant SNPs
}
}

\keyword{internal}
