% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_matrix}
\alias{ld_matrix}
\title{Get LD Matrix}
\usage{
ld_matrix(variant_ids = c())
}
\arguments{
\item{variant_ids}{A character string specifying the Variant ID.  Variant IDs can be SNP IDs or variant IDs.}
}
\value{
A list containing the LD matrix
}
\description{
Get LD matrix from the API by Variant ID
}
\section{ld dataframe}{

The ld dataframe contains information about the LD matrix.
It has the following columns:
\itemize{
\item lead_snp_id: the id of the lead SNP
\item variant_snp_id: the id of the variant SNP
\item ld_block_id: the id of the LD block
\item r: the r value between the lead and variant SNPs
}
}

