% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{ld_matrix_by_snp_id}
\alias{ld_matrix_by_snp_id}
\title{Get LD Matrix by SNP ID}
\usage{
ld_matrix_by_snp_id(snp_ids)
}
\arguments{
\item{snp_ids}{A character string specifying the SNP ID}
}
\value{
A list containing the LD matrix
}
\description{
Get LD matrix from the API by SNP ID
}
