% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{ld_matrix_by_variant}
\alias{ld_matrix_by_variant}
\title{Get LD Matrix by Variant ID}
\usage{
ld_matrix_by_variant(variants)
}
\arguments{
\item{variant_ids}{A character string specifying the Variant ID}
}
\value{
A list containing the LD matrix
}
\description{
Get LD matrix from the API by Variant ID
}
