% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ld_matrix_by_variant_api}
\alias{ld_matrix_by_variant_api}
\title{Get LD Matrix by Variant ID API}
\usage{
ld_matrix_by_variant_api(variants)
}
\arguments{
\item{variants}{A character string specifying the Variant ID}
}
\value{
A list containing the LD matrix
}
\description{
Get LD matrix from the API by Variant ID
}
