% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.R
\name{ld_proxies}
\alias{ld_proxies}
\title{LD Proxies}
\usage{
ld_proxies(variant_ids = c())
}
\arguments{
\item{variant_ids}{A character string specifying the Variant ID.  Variant IDs can be SNP IDs or variant IDs.}
}
\value{
A list containing the LD proxies
}
\description{
Get LD proxies from the API by Variant ID
}
\section{ld_dataframe}{

The ld dataframe contains information about the LD matrix.
It has the following columns:
\itemize{
\item lead_snp_id: the id of the lead SNP
\item variant_snp_id: the id of the variant SNP
\item ld_block_id: the id of the LD block
\item r: the r value between the lead and variant SNPs
}
}

