% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{ld_proxies_by_snp_id_api}
\alias{ld_proxies_by_snp_id_api}
\title{Get LD Proxies by SNP ID API}
\usage{
ld_proxies_by_snp_id_api(snp_ids)
}
\arguments{
\item{snp_ids}{A character string specifying the SNP ID}
}
\value{
A list containing the LD proxies
}
\description{
Get LD proxies from the API by SNP ID
}
