% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{ld_proxies_by_variant}
\alias{ld_proxies_by_variant}
\title{Get LD Proxies by Variant ID}
\usage{
ld_proxies_by_variant(variants)
}
\arguments{
\item{variants}{A character string specifying the Variant ID}
}
\value{
A list containing the LD proxies
}
\description{
Get LD proxies from the API by Variant ID
}
