% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{merge_associations}
\alias{merge_associations}
\title{Merge associations into coloc info}
\usage{
merge_associations(coloc_groups, rare_results, associations)
}
\arguments{
\item{coloc_groups}{A dataframe of coloc groups}

\item{rare_results}{A dataframe of rare results}

\item{associations}{A dataframe of associations}
}
\value{
An updated list of coloc info with associations merged in
}
\description{
Merge associations into coloc info
}
