% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_consistency.R
\name{nds_matrix}
\alias{nds_matrix}
\title{Pairwise Net Directionality Score Matrix}
\usage{
nds_matrix(trait_ids, coloc_groups = NULL, p_threshold = 5e-08)
}
\arguments{
\item{trait_ids}{A numeric vector of trait IDs}

\item{coloc_groups}{A dataframe of coloc_groups with betas (include_associations = TRUE).
If NULL, will fetch via traits(trait_ids, include_associations = TRUE).}

\item{p_threshold}{P-value threshold. Defaults to 5e-8.}
}
\value{
A matrix of NDS values with trait IDs as row/column names.
Diagonal is 1. Values range from -1 (all discordant) to 1 (all concordant).
}
\description{
Compute NDS for all pairs of traits in a set.
}
