% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_consistency.R
\name{pcs_matrix}
\alias{pcs_matrix}
\title{Pairwise Pleiotropic Consistency Score Matrix}
\usage{
pcs_matrix(trait_ids, coloc_groups = NULL, p_threshold = 5e-08)
}
\arguments{
\item{trait_ids}{A numeric vector of trait IDs}

\item{coloc_groups}{A dataframe of coloc_groups covering all requested traits.
If NULL, will fetch via traits(trait_ids).}

\item{p_threshold}{P-value threshold. Defaults to 5e-8.}
}
\value{
A symmetric matrix of PCS values with trait IDs as row/column names.
}
\description{
Compute PCS for all pairs of traits in a set.
}
