% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{rare_results_doc}
\alias{rare_results_doc}
\title{Rare Results Dataframe Documentation}
\usage{
rare_results_doc()
}
\description{
Shared documentation for rare_results dataframe columns
}
\section{rare_results_dataframe}{

The rare_results dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item rare_result_group_id: the unique id for this rare result group
\item study_id: the id of the study associated with this rare result
\item study_extraction_id: the id of the study extraction associated with this rare result
\item snp_id: the id of the SNP
\item ld_block_id: the id of the LD block
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item display_snp: the display SNP name
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
\item ld_block: the LD block of the SNP
}
}

\keyword{internal}
