% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Region}
\usage{
region(
  region_id,
  include_associations = FALSE,
  coloc_group_threshold = "strong",
  include_coloc_pairs = FALSE,
  h4_threshold = 0.8
)
}
\arguments{
\item{region_id}{A numeric value specifying the region id}

\item{include_associations}{A logical value specifying whether to include associations
(BETA, SE, P), defaults to FALSE}

\item{coloc_group_threshold}{A character value specifying the group threshold for coloc groups, defaults to 'strong'}

\item{include_coloc_pairs}{A logical value specifying whether to include coloc pairs, defaults to FALSE}

\item{h4_threshold}{A numeric value specifying the h4 threshold for coloc pairs, defaults to 0.8}
}
\value{
A list which contains the following elements:
\itemize{
\item gene: A list containing metadata about the gene, including region, and neighboring genes.
\item coloc_groups: a dataframe containing information about which studies have coloc results for this gene.
See below for details.
\item study_extractions: a list of dataframes containing the study extractions for this trait.
See below for details.
\item rare_results: (optional) a list of dataframes containing the rare results for this trait
\item coloc_pairs: (optional) a dataframe containing all pairwise coloc results for this trait.
\item variants: a dataframe containing the variants for each associated coloc group or rare group.
}
See below for details.
}
\description{
A collection of studies that are associated with a particular region.
}
\details{
The dataframes returned by this function are as follows:
}
\section{coloc_groups_dataframe}{

The coloc_groups dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item coloc_group_id: the unique id for this group of colocalised results
\item study_id: the id of the study
\item study_extraction_id: the id of the study extraction
\item snp_id: the id of the SNP
\item ld_block_id: the id of the LD block
\item group_threshold: the threshold for this group of colocalised results.
'strong' signifies h4 > 0.8, 'moderate' signifies h4 > 0.6
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item cis_trans: the cis/trans status of the SNP
\item ld_block: the LD block of the SNP
\item display_snp: the display SNP name
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
}
}

\section{genes_in_region_dataframe}{

The genes_in_region dataframe contains information about which genes are in a region.
It has the following columns:
\itemize{
\item id: the id of the gene
\item ensembl_id: the ensembl id of the gene
\item gene: the name of the gene
\item description: the description of the gene
\item gene_biotype: the gene biotype
\item chr: the chromosome of the gene
\item start: the start position of the gene
\item stop: the stop position of the gene
\item strand: the strand of the gene
\item source: the source of the gene
}
}

\section{study_extractions_dataframe}{

The study_extractions dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item id: the unique id for this study extraction
\item study_id: the id of the study associated with this study extraction
\item snp_id: the id of the SNP
\item snp: the SNP name
\item ld_block_id: the id of the LD block
\item unique_study_id: the unique id for this study
\item study: the study name
\item file: the file name
\item svg_file: the SVG file name
\item file_with_lbfs: the file name with lbfs
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item cis_trans: the cis/trans status of the SNP
\item ld_block: the LD block of the SNP
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
}
}

\section{rare_results_dataframe}{

The rare_results dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item rare_result_group_id: the unique id for this rare result group
\item study_id: the id of the study associated with this rare result
\item study_extraction_id: the id of the study extraction associated with this rare result
\item snp_id: the id of the SNP
\item ld_block_id: the id of the LD block
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item display_snp: the display SNP name
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
\item ld_block: the LD block of the SNP
}
}

\section{coloc_pairs_dataframe}{

The coloc_pairs dataframe contains information about which studies have coloc pairs.
It has the following columns:
\itemize{
\item study_extraction_a_id: the id of the study extraction associated with this coloc pair
\item study_extraction_b_id: the id of the study extraction associated with this coloc pair
\item ld_block_id: the id of the LD block
\item h3: the h3 value for this coloc pair
\item h4: the h4 value for this coloc pair
\item spurious: whether this coloc pair is spurious
}
}

\section{variants_dataframe}{

The variants dataframe contains variant information that is pulled from the Variant Effect Predictor (VEP) database.
It has the following columns, along side many more columns from VEP:
\itemize{
\item id: the id of the SNP
\item gene_id: the id of the gene as predicted by VEP
\item gene: the gene name as predicted by VEP
}
}

