% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{region_api}
\alias{region_api}
\title{Get a Region API}
\usage{
region_api(
  region_id,
  include_associations = FALSE,
  include_coloc_pairs = FALSE,
  h4_threshold = 0.8
)
}
\arguments{
\item{region_id}{A character string specifying the region ID}

\item{include_associations}{A logical value specifying whether to include associations}

\item{include_coloc_pairs}{A logical value specifying whether to include coloc pairs}

\item{h4_threshold}{A numeric value specifying the h4 threshold for coloc pairs}
}
\value{
A list containing the region information
}
\description{
Get a region from the API
}
