% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_gpmap}
\alias{search_gpmap}
\title{Search the Genotype-Phenotype Map}
\usage{
search_gpmap(search_text, rsquared_threshold = 0.8)
}
\arguments{
\item{search_text}{A character string specifying the search text}

\item{rsquared_threshold}{A numeric value specifying the rsquared threshold for proxy variants, defaults to 0.8}
}
\value{
A dataframe containing the search results, ordered by the number of colocalisation groups plus rare
results (most first).  For variant searches the searched variant is always first.  Columns include:
\itemize{
\item type: the type of the search result: "original_variant", "proxy_variant", "trait", "gene"
\item name: the name of the search result
\item type_id: the type_id of the search result.  This is the internal id in which the data can be accessed.
\item call: the call to get the search result: "variant(type_id)", "trait(type_id)", "gene(type_id)"
\item info: a string containing informaiton about the search result, which may include:
\itemize{
\item Extractions: the number of extractions
\item Colocalisation Groups: the number of colocalisation groups
\item Colocalisation Studies: the number of colocalisation studies
\item Rare Results: the number of rare results
\item Rsquared: the rsquared of the proxy variant compared to the original variant
}
}
}
\description{
Search the GP Map for Traits, Genes or Variants
}
\details{
After calling search, you can use call the subsequent data as described in the \code{call} column of the
search results.
}
