% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{search_variants_api}
\alias{search_variants_api}
\title{Search Variants API}
\usage{
search_variants_api(query, rsquared_threshold = 0.8)
}
\arguments{
\item{query}{A character string specifying the query}

\item{rsquared_threshold}{A numeric value specifying the rsquared threshold for the proxy variants}
}
\value{
A list containing the variants
}
\description{
Search variants from the API
}
