% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{snp_pleiotropies_api}
\alias{snp_pleiotropies_api}
\title{Get All SNP Pleiotropies API}
\usage{
snp_pleiotropies_api()
}
\value{
A list containing the SNP pleiotropies
}
\description{
Get all SNP pleiotropies from the API
}
