% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{snp_pleiotropies_doc}
\alias{snp_pleiotropies_doc}
\title{SNP Pleiotropies Dataframe Documentation}
\usage{
snp_pleiotropies_doc()
}
\description{
Shared documentation for snp_pleiotropies dataframe columns
}
\section{snp_pleiotropies_dataframe}{

The snp_pleiotropies dataframe contains information about which SNPs are pleiotropic.
It has the following columns:
\itemize{
\item snp_id: the id of the SNP
\item snp: the name of the SNP
\item : distinct_trait_categories the number of trait categories that the SNP is associated with via coloc groups
\item : distinct_protein_coding_genes the number of genes that the SNP is associated with via coloc groups
}
}

\keyword{internal}
