% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{specific_traits_api}
\alias{specific_traits_api}
\title{Get Specific Traits API}
\usage{
specific_traits_api(
  trait_ids,
  include_associations = FALSE,
  include_coloc_pairs = FALSE,
  h4_threshold = 0.8
)
}
\arguments{
\item{trait_ids}{A vector of trait ids}

\item{include_associations}{A logical value specifying whether to include associations}

\item{include_coloc_pairs}{A logical value specifying whether to include coloc pairs}

\item{h4_threshold}{A numeric value specifying the h4 threshold for coloc pairs}
}
\value{
A list containing the traits
}
\description{
Get specific traits from the API
}
