% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{study_extractions_doc}
\alias{study_extractions_doc}
\title{Study Extractions Dataframe Documentation}
\usage{
study_extractions_doc()
}
\description{
Shared documentation for study_extractions dataframe columns
}
\section{study_extractions dataframe}{

The study_extractions dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item id: the unique id for this study extraction
\item study_id: the id of the study associated with this study extraction
\item snp_id: the id of the SNP
\item snp: the SNP name
\item ld_block_id: the id of the LD block
\item unique_study_id: the unique id for this study
\item study: the study name
\item file: the file name
\item svg_file: the SVG file name
\item file_with_lbfs: the file name with lbfs
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item cis_trans: the cis/trans status of the SNP
\item ld_block: the LD block of the SNP
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
}
}

\keyword{internal}
