% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{summary_statistics_doc}
\alias{summary_statistics_doc}
\title{Summary Statistics Dataframe Documentation}
\usage{
summary_statistics_doc()
}
\description{
Shared documentation for summary_statistics dataframe columns
}
\section{summary_statistics_dataframe}{

The summary_statistics dataframe contains information about which studies have summary statistics.
From the API, column names are typically uppercase (SNP, CHR, BP, EA, OA, EAF, Z, BETA, SE, P, LBF_1, etc.).
It has the following columns (names may be upper or lower case depending on source):
\itemize{
\item SNP / snp_id: the id of the SNP
\item CHR / chr: the chromosome of the SNP
\item BP / bp: the base pair position of the SNP
\item EA / ea: the effect allele
\item OA / oa: the other allele
\item EAF / eaf: the estimated allele frequency
\item Z / z: the z-score
\item BETA / beta: the beta value
\item SE / se: the standard error
\item P / p: the p-value
\item imputed: whether the summary statistics are imputed
\item LBF_* / lbf_*: all different finemapped log-bayes factors for each credible set.
Each credible set is numbered from 1 to 10.  If finemapped failed or only returned 1 credible set, the LBF_1
column is just converted directly from the z-score.
}
}

\keyword{internal}
