% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{trait}
\alias{trait}
\title{Get a Trait}
\usage{
trait(trait_id, include_associations = FALSE)
}
\arguments{
\item{trait_id}{A character string specifying the trait ID}

\item{include_associations}{A logical value specifying whether to include associations}
}
\value{
A list containing the trait
}
\description{
Get a trait from the API
}
