% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{trait_api}
\alias{trait_api}
\title{Get a Trait API}
\usage{
trait_api(
  trait_id,
  include_associations = FALSE,
  include_coloc_pairs = FALSE,
  h4_threshold = 0.8
)
}
\arguments{
\item{trait_id}{A character string specifying the trait ID}

\item{include_associations}{A logical value specifying whether to include associations}

\item{include_coloc_pairs}{A logical value specifying whether to include coloc pairs}

\item{h4_threshold}{A numeric value specifying the h4 threshold for coloc pairs}
}
\value{
A list containing the trait
}
\description{
Get a trait from the API
}
