% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{trait_duplicates}
\alias{trait_duplicates}
\title{Trait duplicates}
\usage{
trait_duplicates()
}
\value{
A dataframe containing the duplicated traits with the following columns:
\itemize{
\item trait_id: the id of the duplicated trait
\item trait_name: the name of the duplicated trait
\item parent_trait_id: the id of the trait it duplicates
\item parent_trait_name: the name of the trait it duplicates
}
}
\description{
Get all traits that have been marked as a duplicate of another trait
}
