% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{upload_gwas}
\alias{upload_gwas}
\title{Upload a GWAS to the API}
\usage{
upload_gwas(
  file,
  name,
  p_value_threshold = 5e-08,
  column_names = NULL,
  email = NA,
  category = "continuous",
  is_published = FALSE,
  doi = NA,
  should_be_added = FALSE,
  ancestry = "EUR",
  sample_size = NA,
  reference_build = "GRCh38",
  compare_with_upload_guids = NA,
  chr_col = NULL,
  bp_col = NULL,
  ea_col = NULL,
  oa_col = NULL,
  p_col = NULL,
  eaf_col = NULL,
  beta_col = NULL,
  se_col = NULL,
  or_col = NULL,
  or_lb_col = NULL,
  or_ub_col = NULL,
  snp_col = NULL,
  rsid_col = NULL
)
}
\arguments{
\item{file}{The path to the GWAS file, maximum size is 1GB}

\item{name}{The name of the GWAS}

\item{p_value_threshold}{The p-value threshold for the GWAS}

\item{column_names}{Deprecated, use the \verb{*_col} arguments instead.
A named list of column names in the format of: list(CHR = "chr", BP = "pos"...).
Accepted names are SNP, RSID, CHR, BP, EA, OA, P, EAF, BETA, SE, OR, OR_LB (or LB) and OR_UB (or UB).}

\item{email}{The email of the user}

\item{category}{The category of the GWAS.  Only "continuous" and "categorical" are accepted.}

\item{is_published}{Whether the GWAS is published}

\item{doi}{The DOI of the GWAS}

\item{should_be_added}{Whether the GWAS should be added to the API}

\item{ancestry}{The ancestry of the GWAS.  Currently only "EUR" is accepted.}

\item{sample_size}{The sample size of the GWAS}

\item{reference_build}{The reference build of the GWAS.  Only "GRCh37" and "GRCh38" are accepted.}

\item{compare_with_upload_guids}{A vector of GUIDs of uploads to compare with}

\item{chr_col}{The name of the chromosome column (required)}

\item{bp_col}{The name of the base pair position column (required)}

\item{ea_col}{The name of the effect allele column (required)}

\item{oa_col}{The name of the other allele column (required)}

\item{p_col}{The name of the p-value column (required)}

\item{eaf_col}{The name of the effect allele frequency column (optional)}

\item{beta_col}{The name of the beta column (required with \code{se_col}, unless using odds ratios)}

\item{se_col}{The name of the standard error column (required with \code{beta_col}, unless using odds ratios)}

\item{or_col}{The name of the odds ratio column (required with \code{or_lb_col} and \code{or_ub_col}, unless using beta)}

\item{or_lb_col}{The name of the odds ratio lower confidence bound column}

\item{or_ub_col}{The name of the odds ratio upper confidence bound column}

\item{snp_col}{The name of the SNP identifier column (optional)}

\item{rsid_col}{The name of the rsID column (optional)}
}
\value{
A list containing the GWAS information, including the \code{guid} used to fetch results
}
\description{
Upload a GWAS to the API.  Column names are specified with the \verb{*_col} arguments, e.g.
\code{chr_col = "chr"}.  \code{chr_col}, \code{bp_col}, \code{ea_col}, \code{oa_col} and \code{p_col} are required, along with either
\code{beta_col} and \code{se_col}, or \code{or_col}, \code{or_lb_col} and \code{or_ub_col}.
}
