% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{upload_gwas_api}
\alias{upload_gwas_api}
\title{Upload a GWAS to the API}
\usage{
upload_gwas_api(
  file,
  name,
  p_value_threshold,
  column_names,
  email,
  category,
  is_published = FALSE,
  doi = NA,
  should_be_added = FALSE,
  ancestry = "EUR",
  sample_size,
  reference_build = "GRCh38"
)
}
\arguments{
\item{file}{The path to the GWAS file, maximum size is 1GB}

\item{name}{The name of the GWAS}

\item{p_value_threshold}{The p-value threshold for the GWAS}

\item{column_names}{A list of column names in the format of: list(CHR = "chr", BP = "pos"...)}

\item{email}{The email of the user}

\item{category}{The category of the GWAS.  Only "continuous" and "categorical" are accepted.}

\item{is_published}{Whether the GWAS is published}

\item{doi}{The DOI of the GWAS}

\item{should_be_added}{Whether the GWAS should be added to the API}

\item{ancestry}{The ancestry of the GWAS.  Currently only "EUR" is accepted.}

\item{sample_size}{The sample size of the GWAS}

\item{reference_build}{The reference build of the GWAS.  Only "GRCh37" and "GRCh38" are accepted.}
}
\value{
A list containing the GWAS information
}
\description{
Upload a GWAS to the API
}
