% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{variant}
\alias{variant}
\title{Get a Variant}
\usage{
variant(snp_id)
}
\arguments{
\item{snp_id}{A character string specifying the SNP ID}
}
\value{
A list containing the variant
}
\description{
Get a variant from the API
}
