% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{variant_pleiotropies_doc}
\alias{variant_pleiotropies_doc}
\title{SNP Pleiotropies Dataframe Documentation}
\usage{
variant_pleiotropies_doc()
}
\description{
Shared documentation for variant_pleiotropies dataframe columns
}
\section{variant_pleiotropies_dataframe}{

The variant_pleiotropies dataframe contains information about which SNPs are pleiotropic.
It has the following columns:
\itemize{
\item variant_id: the id of the SNP
\item snp: the name of the SNP
\item : distinct_trait_categories the number of trait categories that the SNP is associated with via coloc groups
\item : distinct_protein_coding_genes the number of genes that the SNP is associated with via coloc groups
}
}

\keyword{internal}
