% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{variant_summary_stats_api}
\alias{variant_summary_stats_api}
\title{Get Variant Summary Stats API}
\usage{
variant_summary_stats_api(snp_id)
}
\arguments{
\item{snp_id}{A character string specifying the SNP ID}
}
\value{
A list containing dataframes from the TSV files
}
\description{
Get variant summary statistics from the API
}
