% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{variants_by_grange_api}
\alias{variants_by_grange_api}
\title{Get Variants by GRange}
\usage{
variants_by_grange_api(
  chr,
  start,
  stop,
  include_associations = FALSE,
  p_value_threshold = NULL
)
}
\arguments{
\item{chr}{A character string specifying the chromosome}

\item{start}{A numeric value specifying the start position}

\item{stop}{A numeric value specifying the stop position}

\item{include_associations}{A logical value specifying whether to include associations}

\item{p_value_threshold}{A numeric value specifying the p-value threshold}
}
\value{
A list containing the variants
}
\description{
Get variants from the API by GRange
}
