% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{variants_by_snp_id_api}
\alias{variants_by_snp_id_api}
\title{Get Variants by SNP ID}
\usage{
variants_by_snp_id_api(
  snp_ids,
  include_associations = FALSE,
  p_value_threshold = NULL
)
}
\arguments{
\item{snp_ids}{A character string specifying the SNP ID}

\item{include_associations}{A logical value specifying whether to include associations}

\item{p_value_threshold}{A numeric value specifying the p-value threshold}
}
\value{
A list containing the variants
}
\description{
Get variants from the API by SNP ID
}
