% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{variants_by_variant}
\alias{variants_by_variant}
\title{Get Variants by Variant ID}
\usage{
variants_by_variant(
  variants,
  include_associations = FALSE,
  p_value_threshold = NULL
)
}
\arguments{
\item{include_associations}{A logical value specifying whether to include associations}

\item{p_value_threshold}{A numeric value specifying the p-value threshold}

\item{variant_ids}{A character string specifying the Variant ID}
}
\value{
A list containing the variants
}
\description{
Get variants from the API by Variant ID
}
