% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{variants_doc}
\alias{variants_doc}
\title{Variants Dataframe Documentation}
\usage{
variants_doc()
}
\description{
Shared documentation for variants dataframe columns
}
\section{variants_dataframe}{

The variants dataframe contains variant information that is pulled from the Variant Effect Predictor (VEP) database.
It has the following columns, along side many more columns from VEP:
\itemize{
\item id: the id of the SNP
\item gene_id: the id of the gene as predicted by VEP
\item gene: the gene name as predicted by VEP
}
}

\keyword{internal}
