% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{variants_grange_api}
\alias{variants_grange_api}
\title{Get Variants by GRange API}
\usage{
variants_grange_api(
  chr,
  start,
  stop,
  include_associations = FALSE,
  include_coloc_pairs = FALSE,
  h4_threshold = 0.8
)
}
\arguments{
\item{chr}{A character string specifying the chromosome}

\item{start}{A numeric value specifying the start position}

\item{stop}{A numeric value specifying the stop position}

\item{include_associations}{Logical. Whether to include associations (only when expand=TRUE, but expand not available with grange)}

\item{include_coloc_pairs}{Logical. Whether to include coloc pairs}

\item{h4_threshold}{Numeric. H4 threshold for coloc pairs, defaults to 0.8}
}
\value{
A list containing the variants
}
\description{
Get variants from the API by genomic range. Expand is not available with grange.
}
