% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{genes}
\alias{genes}
\title{Get Gene Information}
\usage{
genes(symbol)
}
\arguments{
\item{symbol}{A character string specifying the gene symbol}
}
\value{
A list containing the gene information
}
\description{
Get gene information from the API
}
