% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiotropy.R
\name{get_all_gene_pleiotropies}
\alias{get_all_gene_pleiotropies}
\title{Get All Gene Pleiotropies}
\usage{
get_all_gene_pleiotropies()
}
\value{
A list containing the gene pleiotropy
}
\description{
Get gene pleiotropy from the API by gene id
}
