% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{merge_associations}
\alias{merge_associations}
\title{Merge associations into coloc info}
\usage{
merge_associations(coloc_info)
}
\arguments{
\item{coloc_info}{A list of coloc info, as returned by the api}
}
\value{
An updated list of coloc info with associations merged in
}
\description{
Merge associations into coloc info
}
