% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{select_api}
\alias{select_api}
\title{Select API}
\usage{
select_api(api = "dev")
}
\arguments{
\item{api}{A character string specifying the API to use. Default is "production".}
}
\description{
Select the API to use
}
