% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{summary_statistics_doc}
\alias{summary_statistics_doc}
\title{Summary Statistics Dataframe Documentation}
\usage{
summary_statistics_doc()
}
\description{
Shared documentation for summary_statistics dataframe columns
}
\section{summary_statistics_dataframe}{

The summary_statistics dataframe contains information about which studies have summary statistics.
It has the following columns:
\itemize{
\item snp_id: the id of the SNP
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item ea: the effect allele
\item oa: the other allele
\item eaf: the estimated allele frequency
\item z: the z-score
\item beta: the beta value
\item se: the standard error
\item p: the p-value
\item imputed: whether the summary statistics are imputed
\item lbf_*: all different finemapped log-bayes factors for each credible set.
Each credible set is numbered from 1 to 10.  If finemapped failed or only returned 1 credible set, the lbf_1
column is just converted directly from the z-score.
}
}

\keyword{internal}
