% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{upload_gwas}
\alias{upload_gwas}
\title{Upload a GWAS to the API}
\usage{
upload_gwas(
  file,
  name,
  p_value_threshold = 5e-08,
  column_names = list(),
  email = NA,
  category = "continuous",
  is_published = FALSE,
  doi = NA,
  should_be_added = FALSE,
  ancestry = "EUR",
  sample_size = NA,
  reference_build = "GRCh38"
)
}
\arguments{
\item{file}{The path to the GWAS file, maximum size is 1GB}

\item{name}{The name of the GWAS}

\item{p_value_threshold}{The p-value threshold for the GWAS}

\item{column_names}{A list of column names in the format of: list(CHR = "chr", BP = "pos"...)
\itemize{
\item CHR: chromosome
\item BP: base pair position
\item P: p-value
\item EA: allele 1
\item OA: allele 2
\item EAF: allele frequency
And either BETA and SE, or OR, LB, and UB
\item BETA: beta
\item SE: standard error
\item OR: odds ratio
\item LB: lower bound of the confidence interval
\item UB: upper bound of the confidence interval
}}

\item{email}{The email of the user}

\item{category}{The category of the GWAS.  Only "continuous" and "categorical" are accepted.}

\item{is_published}{Whether the GWAS is published}

\item{doi}{The DOI of the GWAS}

\item{should_be_added}{Whether the GWAS should be added to the API}

\item{ancestry}{The ancestry of the GWAS.  Currently only "EUR" is accepted.}

\item{sample_size}{The sample size of the GWAS}

\item{reference_build}{The reference build of the GWAS.  Only "GRCh37" and "GRCh38" are accepted.}
}
\value{
A list containing the GWAS information
}
\description{
Upload a GWAS to the API
}
