% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variant}
\alias{variant}
\title{Variant}
\usage{
variant(
  snp_id,
  coloc_group_threshold = "strong",
  include_coloc_pairs = FALSE,
  h4_threshold = 0.8,
  include_summary_stats = FALSE
)
}
\arguments{
\item{snp_id}{A character string specifying the SNP ID}

\item{coloc_group_threshold}{A character value specifying the group threshold for coloc groups, defaults to 'strong'}

\item{include_coloc_pairs}{A logical value specifying whether to include coloc pairs}

\item{h4_threshold}{A numeric value specifying the cutoff for included coloc pairs, defaults to 0.8.
Only used if include_coloc_pairs is TRUE.}

\item{include_summary_stats}{A logical value specifying whether to include summary stats}
}
\value{
A list which contains the following elements:
\itemize{
\item variant: named list containing the variant information
\item coloc_groups: a dataframe containing information about which studies have coloc results for this variant
\item rare_results: a list of dataframes containing the rare variants
\item study_extractions: a list of dataframes containing the study extractions
\item summary_stats (optional): a list of dataframes containing the summary stats for each study,
where the name of the dataframe is the study_extraction_id
\item coloc_pairs (optional): a dataframe containing information about which studies have coloc pairs for this variant
where the study_extraction_a_id and study_extraction_b_id are the study_extraction_ids of the two studies.
h4_threshold is the cutoff for included coloc pairs, defaults to 0.8
}
}
\description{
A collection of studies that are associated with a particular variant.
}
