% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{variant}
\alias{variant}
\title{Variant}
\usage{
variant(
  snp_id,
  coloc_group_threshold = "strong",
  include_coloc_pairs = FALSE,
  h4_threshold = 0.8,
  include_summary_stats = FALSE
)
}
\arguments{
\item{snp_id}{A character string specifying the SNP ID}

\item{coloc_group_threshold}{A character value specifying the group threshold for coloc groups, defaults to 'strong'}

\item{include_coloc_pairs}{A logical value specifying whether to include coloc pairs}

\item{h4_threshold}{A numeric value specifying the cutoff for included coloc pairs, defaults to 0.8.
Only used if include_coloc_pairs is TRUE.}

\item{include_summary_stats}{A logical value specifying whether to include summary stats}
}
\value{
A list which contains the following elements:
\itemize{
\item variant: named list containing the variant information
\item coloc_groups: a dataframe containing information about which studies have coloc results for this variant
\item rare_results: a list of dataframes containing the rare variants
\item study_extractions: a list of dataframes containing the study extractions
\item summary_stats (optional): a list of dataframes containing the summary stats for each study,
where the name of the dataframe is the study_extraction_id
\item coloc_pairs (optional): a dataframe containing information about which studies have coloc pairs for this variant
where the study_extraction_a_id and study_extraction_b_id are the study_extraction_ids of the two studies.
h4_threshold is the cutoff for included coloc pairs, defaults to 0.8
}
}
\description{
A collection of studies that are associated with a particular variant.
}
\details{
The dataframes returned by this function are as follows:
}
\section{coloc_groups_dataframe}{

The coloc_groups dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item coloc_group_id: the unique id for this group of colocalised results
\item study_id: the id of the study
\item study_extraction_id: the id of the study extraction
\item snp_id: the id of the SNP
\item ld_block_id: the id of the LD block
\item group_threshold: the threshold for this group of colocalised results.
'strong' signifies h4 > 0.8, 'moderate' signifies h4 > 0.6
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item cis_trans: the cis/trans status of the SNP
\item ld_block: the LD block of the SNP
\item display_snp: the display SNP name
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
}
}

\section{rare_results_dataframe}{

The rare_results dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item rare_result_group_id: the unique id for this rare result group
\item study_id: the id of the study associated with this rare result
\item study_extraction_id: the id of the study extraction associated with this rare result
\item snp_id: the id of the SNP
\item ld_block_id: the id of the LD block
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item display_snp: the display SNP name
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
\item ld_block: the LD block of the SNP
}
}

\section{study_extractions_dataframe}{

The study_extractions dataframe contains information about which studies have coloc results.
It has the following columns:
\itemize{
\item id: the unique id for this study extraction
\item study_id: the id of the study associated with this study extraction
\item snp_id: the id of the SNP
\item snp: the SNP name
\item ld_block_id: the id of the LD block
\item unique_study_id: the unique id for this study
\item study: the study name
\item file: the file name
\item svg_file: the SVG file name
\item file_with_lbfs: the file name with lbfs
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item min_p: the minimum p-value related to the study_extraction_id
\item cis_trans: the cis/trans status of the SNP
\item ld_block: the LD block of the SNP
\item gene: the gene associated with the SNP
\item gene_id: the id of the gene
\item trait_id: the id of the trait
\item trait_name: the name of the trait
\item trait_category: the category of the trait
\item data_type: the data type of the trait
\item tissue: the tissue of the trait
}
}

\section{summary_statistics_dataframe}{

The summary_statistics dataframe contains information about which studies have summary statistics.
It has the following columns:
\itemize{
\item snp_id: the id of the SNP
\item chr: the chromosome of the SNP
\item bp: the base pair position of the SNP
\item ea: the effect allele
\item oa: the other allele
\item eaf: the estimated allele frequency
\item z: the z-score
\item beta: the beta value
\item se: the standard error
\item p: the p-value
\item imputed: whether the summary statistics are imputed
\item lbf_*: all different finemapped log-bayes factors for each credible set.
Each credible set is numbered from 1 to 10.  If finemapped failed or only returned 1 credible set, the lbf_1
column is just converted directly from the z-score.
}
}

\section{coloc_pairs_dataframe}{

The coloc_pairs dataframe contains information about which studies have coloc pairs.
It has the following columns:
\itemize{
\item study_extraction_a_id: the id of the study extraction associated with this coloc pair
\item study_extraction_b_id: the id of the study extraction associated with this coloc pair
\item ld_block_id: the id of the LD block
\item h3: the h3 value for this coloc pair
\item h4: the h4 value for this coloc pair
\item spurious: whether this coloc pair is spurious
}
}

