% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{variants_by_rsid_api}
\alias{variants_by_rsid_api}
\title{Get Variants by RSID}
\usage{
variants_by_rsid_api(rsids)
}
\arguments{
\item{rsids}{A character string specifying the RSID}
}
\value{
A list containing the variants
}
\description{
Get variants from the API by RSID
}
